refactor regression datasets
