+++ pyrelational/datasets/regression/__init__.py
@@ -1,0 +1,10 @@
+from pyrelational.datasets.regression.scikit_learn import DiabetesDataset
+from pyrelational.datasets.regression.synthetic import SynthReg1, SynthReg2
+from pyrelational.datasets.regression.uci import (
+    UCIAirfoil,
+    UCIConcrete,
+    UCIEnergy,
+    UCIPower,
+    UCIWine,
+    UCIYacht,
+)
